TS: TS_READ32_COMMAND, special RO pages for FS image

+++ firmware/console/binary/tunerstudio_impl.h
@@ -34,7 +34,7 @@ public:
 	void handleWriteChunkCommand(TsChannelBase* tsChannel, uint16_t page, uint16_t offset, uint16_t count,
 			void *content);
 	void handleCrc32Check(TsChannelBase *tsChannel,  uint16_t page, uint16_t offset, uint16_t count);
-	void handlePageReadCommand(TsChannelBase* tsChannel, uint16_t page, uint16_t offset, uint16_t count);
+	void handlePageReadCommand(TsChannelBase* tsChannel, uint32_t page, uint32_t offset, uint32_t count);
 	void handleScatteredReadCommand(TsChannelBase* tsChannel);
 
 private:
